eco_pro/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
backcall==0.2.0
braintree==4.18.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==22.1
django-heroku==0.3.1
django-polymorphic==3.1.0
django-rest-knox==4.2.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
drf-yasg2==1.19.4
dynaconf==3.1.11
executing==1.2.0
html5lib==1.1
hvac==1.0.2
idna==3.4
inflection==0.5.1
ipdb==0.13.11
ipython==8.11.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
knox==0.1.14
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
numpy==1.24.2
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyHanko==0.17.0
pyhanko-certvalidator==0.20.1
pyhcl==0.4.4
PyJWT==1.7.1
pypdf==3.4.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
shortuuid==1.0.11
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.6.2
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
validators==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.4.0
win32-setctime==1.1.0
xhtml2pdf==0.2.9